Move remaining helper functions to lib/
- Clear the regular update cache between unit tests
