+++ lib/in-process-server-test-helpers.js
@@ -45,7 +45,7 @@ function start () {
  * @param {Object} server instance
  */
 function reset (server) {
-  server.requestCache.clear();
+  server.reset();
 }
 
 /**
